+++ JL.Core/Dicts/EPWING/Yomichan/EpwingYomichanRecord.cs
@@ -15,7 +15,7 @@ internal sealed class EpwingYomichanRecord : IDictRecord, IGetFrequency, IEquata
 {
     public string PrimarySpelling { get; }
     public string? Reading { get; }
-    public int PopularityScore { get; }
+    public double PopularityScore { get; }
     public string[]? Definitions { get; }
     public string[]? WordClasses { get; }
     public string[]? DefinitionTags { get; }
@@ -23,7 +23,7 @@ internal sealed class EpwingYomichanRecord : IDictRecord, IGetFrequency, IEquata
     //public int Sequence { get; }
     //public string[]? TermTags { get; }
 
-    public EpwingYomichanRecord(string primarySpelling, string? reading, int popularityScore, string[]? definitions, string[]? wordClasses, string[]? definitionTags, ImageInfo[]? imageInfo)
+    public EpwingYomichanRecord(string primarySpelling, string? reading, double popularityScore, string[]? definitions, string[]? wordClasses, string[]? definitionTags, ImageInfo[]? imageInfo)
     {
         PrimarySpelling = primarySpelling;
         Reading = reading;
